refactor: proper conversion between integer types (#148)

+++ internal/serialmonitor/capture.go
@@ -1,11 +1,11 @@
 package serialmonitor
 
 type capture struct {
-	elapsedTime      uint64
-	distance         uint64
-	time500m         uint64
-	strokesPerMinute uint64
-	watts            uint64
-	cals             uint64
-	level            uint64
+	elapsedTime      uint16
+	distance         uint16
+	time500m         uint16
+	strokesPerMinute uint8
+	watts            uint16
+	cals             uint16
+	level            uint8
 }
